Fix signedness warnings in tests

+++ test/relative_test.c
@@ -11,7 +11,7 @@ int relative_check()
     "./simple", ".././simple"};
   const char *absolute_paths[] = {"/", "/test", "/../test/", "/../another_test",
     "/./simple", "/.././simple"};
-  int i;
+  size_t i;
 
   cwk_path_set_style(CWK_STYLE_UNIX);
